+++ app/src/main/java/tech/caaa/aircraft/game/UserInput.kt
@@ -1,6 +1,6 @@
 package tech.caaa.aircraft.game
 
 sealed class UserInput {
-    data class MovePlane(val planeId: UInt, val x: Float, val y: Float) : UserInput()
+    data class MovePlane(val planeId: UInt, val x: Double, val y: Double) : UserInput()
     data object HoldShoot: UserInput()
 }
